+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.0-rc.3] - Unreleased
+
+## Added
+
+- `rivet global-config read-project` command
+- `rivet global-config path` command to get the path to the global config
+
+## Fixed
+
+- `rivet unlink` now works even if the credentials are invalid
+
 ## [v1.0.0-rc.2] - 2024-01-13
 
 ## Added